gather2-core/requirements.txt
################################################################################
#                                                                              #
#                              Frozen requirements                             #
#                                                                              #
#                               DO NOT CHANGE !!!                              #
#                                                                              #
# To update this first update 'primary-requirements.txt'                       #
# then run 'docker-compose run core pip_freeze'                                #
#                                                                              #
################################################################################

certifi==2017.4.17
chardet==3.0.4
contextlib2==0.5.5
coverage==4.4.1
decorator==4.1.2
Django==1.11.3
django-attachments==1.1.1
django-cas-ng==3.5.8
django-cors-headers==2.1.0
django-extensions==1.8.1
django-filter==1.0.4
django-reversion==2.0.9
django-reversion-compare==0.7.5
django-ums-client==0.2.3
djangorestframework==3.6.3
djangorestframework-filters==0.10.1
drf-extensions==0.3.1
flake8==3.3.0
hypothesis==3.14.0
idna==2.5
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jsonschema==2.6.0
Markdown==2.6.8
mccabe==0.6.1
mock==2.0.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
python-cas==1.2.0
pytz==2017.2
raven==6.1.0
requests==2.18.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
urllib3==1.22
uWSGI==2.0.15
wcwidth==0.1.7
